#456 fixed bitrate not being saved
